MAT-6631: Remove unneeded util, move to existing util
